+++ src/main/java/domain/Collection/Results.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.Collection;
 
 import java.util.Arrays;
 import java.util.Collections;
