+++ src/main/java/domain/LadderGame/SequenceSwapper.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.LadderGame;
 
 import java.util.ArrayList;
 import java.util.Arrays;
