+++ src/main/java/view/LadderView.java
@@ -1,7 +1,7 @@
 package view;
 
-import domain.Ladder;
-import domain.Line;
+import domain.Ladder.Ladder;
+import domain.Ladder.Line;
 import domain.util.Point;
 
 import java.util.stream.Collectors;
