+++ src/test/java/domain/ParticipantsTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.Collection.Participant;
+import domain.Collection.Participants;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
