+++ src/test/java/domain/ResultsTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.Collection.Result;
+import domain.Collection.Results;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
