+++ src/test/java/domain/SequenceSwapperTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.LadderGame.SequenceSwapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
